+++ GlobalUsings.cs
@@ -21,6 +21,8 @@ global using BepInEx.Configuration;
 global using BepInEx.Logging;
 global using System.IO;
 global using UnityEngine;
+
 global using static WECCL.Utils.TextureUtils;
 global using static WECCL.Utils.MenuUtils;
 global using static WECCL.Content.CustomContent;
+global using static WECCL.Utils.FileNameUtils;