+++ Saves/ContentMappings.cs
@@ -1,4 +1,5 @@
 ﻿using Newtonsoft.Json;
+using WECCL.Utils;
 
 namespace WECCL.Saves;
 
@@ -10,31 +11,31 @@ internal class ContentMappings
     {
         for (int i = 0; i < 40; i++)
         {
-            this.MaterialNameMap.Add(new List<string>());
-            this.FleshNameMap.Add(new List<string>());
-            this.ShapeNameMap.Add(new List<string>());
+            this.MaterialNameMap.Add(new ContentList());
+            this.FleshNameMap.Add(new ContentList());
+            this.ShapeNameMap.Add(new ContentList());
         }
     }
 
     internal static ContentMappings ContentMap => _instance ??= new ContentMappings();
 
-    public List<List<string>> MaterialNameMap { get; set; } = new();
-    public List<List<string>> FleshNameMap { get; set; } = new();
-    public List<List<string>> ShapeNameMap { get; set; } = new();
+    public List<ContentList> MaterialNameMap { get; set; } = new();
+    public List<ContentList> FleshNameMap { get; set; } = new();
+    public List<ContentList> ShapeNameMap { get; set; } = new();
 
-    public List<string> FaceFemaleNameMap { get; set; } = new();
+    public ContentList FaceFemaleNameMap { get; set; } = new();
 
-    public List<string> BodyFemaleNameMap { get; set; } = new();
+    public ContentList BodyFemaleNameMap { get; set; } = new();
 
-    public List<string> SpecialFootwearNameMap { get; set; } = new();
+    public ContentList SpecialFootwearNameMap { get; set; } = new();
 
-    public List<string> TransparentHairMaterialNameMap { get; set; } = new();
+    public ContentList TransparentHairMaterialNameMap { get; set; } = new();
 
-    public List<string> TransparentHairHairstyleNameMap { get; set; } = new();
+    public ContentList TransparentHairHairstyleNameMap { get; set; } = new();
 
-    public List<string> KneepadNameMap { get; set; } = new();
+    public ContentList KneepadNameMap { get; set; } = new();
 
-    public List<string> MusicNameMap { get; set; } = new();
+    public ContentList MusicNameMap { get; set; } = new();
     
     public List<string> PreviouslyImportedCharacters { get; set; } = new();