+++ Utils/FileNameUtils.cs
@@ -1,6 +1,6 @@
 ﻿namespace WECCL.Utils;
 
-public class FileNameUtils
+public static class FileNameUtils
 {
     public static string Escape(string fileName)
     {
@@ -10,5 +10,18 @@ public class FileNameUtils
             fileName = fileName.Replace(c, '_');
         }
         return fileName;
+    }
+
+    public static string FindPluginName(DirectoryInfo source)
+    {
+        while (source.Name != "plugins" && source.Parent != null)
+        {
+            source = source.Parent;
+        }
+        if (source.Parent?.Name == "BepInEx")
+        {
+            return "Assets";
+        }
+        return source.Parent?.Name ?? "Assets";
     }
 }